pyproject.toml
[project]
name = "simpleagent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain-google-genai>=2.1.5",
    "langchain-mcp-adapters>=0.1.7",
    "langgraph>=0.4.8",
    "python-dotenv>=1.1.0",
]
